refactor: use a single spring configuration for infra services
