+++ gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/src/main/java/io/gravitee/rest/api/management/v2/rest/spring/RestManagementConfiguration.java
@@ -15,8 +15,7 @@
  */
 package io.gravitee.rest.api.management.v2.rest.spring;
 
-import io.gravitee.apim.infra.UsecaseSpringConfiguration;
-import io.gravitee.apim.infra.crud_service.CrudServiceSpringConfiguration;
+import io.gravitee.apim.infra.spring.UsecaseSpringConfiguration;
 import io.gravitee.el.ExpressionLanguageInitializer;
 import io.gravitee.plugin.core.spring.PluginConfiguration;
 import io.gravitee.rest.api.idp.core.spring.IdentityProviderPluginConfiguration;
@@ -31,13 +30,7 @@ import org.springframework.context.annotation.Import;
  */
 @Configuration
 @Import(
-    {
-        PluginConfiguration.class,
-        ServiceConfiguration.class,
-        IdentityProviderPluginConfiguration.class,
-        UsecaseSpringConfiguration.class,
-        CrudServiceSpringConfiguration.class,
-    }
+    { PluginConfiguration.class, ServiceConfiguration.class, IdentityProviderPluginConfiguration.class, UsecaseSpringConfiguration.class }
 )
 public class RestManagementConfiguration {
