+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/apim/infra/spring/InfraServiceSpringConfiguration.java
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.apim.infra.crud_service;
+package io.gravitee.apim.infra.spring;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@ComponentScan(basePackages = { "io.gravitee.apim.infra.crud_service" })
-public class CrudServiceSpringConfiguration {}
+@ComponentScan(basePackages = { "io.gravitee.apim.infra" })
+public class InfraServiceSpringConfiguration {}
